fixes failing tests being running from under an unpriviledged user
accesses to proc items require extra permission. execute these
actions in sudo subshell.
fallback to the original even-odd approach if the proc entry for
ip_local_reserved_ports is unavailable which is the case for some
exotic test environments.

related: CEM-15664

+++ src/config/common/cfgm_common/tests/test_utils.py
@@ -776,6 +776,58 @@ fake_keystone_client = FakeKeystoneClient()
 def get_keystone_client(*args, **kwargs):
     return fake_keystone_client
 
+#
+# Find two consecutive free ports such that even port is greater than odd port
+# Return the even port and socket locked to the odd port
+def _get_free_even_port(allocated_sockets):
+    single_port_list = []
+    tmp_sock2 = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    while (1):
+        tmp_sock1 = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        try:
+            tmp_sock1.bind(('', 0))
+        except:
+            #Bail out as all ports exhausted.
+            for tmp_sock in single_port_list:
+                tmp_sock.close()
+            raise
+            return None, 0
+
+        free_port1 = tmp_sock1.getsockname()[1]
+        if free_port1 % 2:
+            #We have odd port, check whether next even port free
+            try:
+                tmp_sock2.bind(('', free_port1 + 1))
+            except:
+                single_port_list.append(tmp_sock1)
+                continue
+        else:
+            #We have even port, check whether next odd port free
+            try:
+                tmp_sock2.bind(('', free_port1 - 1))
+            except:
+                single_port_list.append(tmp_sock1)
+                continue
+        free_port2 = tmp_sock2.getsockname()[1]
+        break
+
+    #we have found our twin ports, release the singles now
+    for tmp_sock in single_port_list:
+        tmp_sock.close()
+
+    #keep the odd port locked and return free port
+    if free_port1 % 2:
+        odd_port, odd_sock = free_port1, tmp_sock1
+        even_port, even_sock = free_port2, tmp_sock2
+    else:
+        odd_port, odd_sock = free_port2, tmp_sock2
+        even_port, even_sock = free_port1, tmp_sock1
+
+    even_sock.close()
+    allocated_sockets.append(odd_sock)
+    return even_port
+
+# end _get_free_even_port
 
 #
 # Finds and reserves a free tcp port
@@ -783,24 +835,35 @@ def get_keystone_client(*args, **kwargs):
 free_port_guard = threading.Lock()
 def get_free_port(allocated_sockets):
     output = 0
-    P = '/proc/sys/net/ipv4/ip_local_reserved_ports'
+    x = 'ip_local_reserved_ports'
+    P = '/proc/sys/net/ipv4/{0}'.format(x)
+    if not os.path.exists(P):
+        return _get_free_even_port(allocated_sockets)
+
     with contextlib.closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
         s.bind(('', 0))
         output = s.getsockname()[1]
         with free_port_guard:
-            with open(P, 'w+') as f:
-                fcntl.lockf(f, fcntl.LOCK_EX)
-                t = str(f.read(2**20)).strip()
-                if 0 < len(t):
-                    t += ','
-                t += str(output)
-                print('TO WRITE: "{0}"'.format(t))
-                F, p = tempfile.mkstemp()
-                with os.fdopen(F, 'w') as T:
-                    T.write(t)
+            F, p = tempfile.mkstemp()
+            with os.fdopen(F, 'w') as T:
+                g = '{0}/{1}.lck'.format(os.path.dirname(p, x))
+                with open(g, 'w+') as G:
+                    fcntl.lockf(G, fcntl.LOCK_EX)
+                    with open(P, 'r') as f:
+                        t = str(f.read(2**20)).strip()
 
-                os.system('cat {0} > {1}'.format(p, P))
-                os.remove(p)
+                    if 0 < len(t):
+                        t += ','
+
+                    t += str(output)
+                    print('TO WRITE: "{0}"'.format(t))
+                    T.write(t)
+                    T.close()
+
+                    e = os.system("sudo -n sh -c 'cat {0} > {1}'".format(p, P))
+                    os.remove(p)
+                    if 0 != e:
+                        raise Exception("Cannot write to %s: %d" % (P, e))
 
     return output
